Move form builder specific bits into a separate component
